+++ pool_tlru.go
@@ -12,11 +12,11 @@ import (
 	"github.com/justpranavrs/tlru/mux"
 )
 
-// TLRU is the better implementation of [lrucore.TTLCore]. It creates
-// many instances of [lrucore.TTLCore] and works based on [LRU].
+// PoolTLRU is the better implementation of [lrucore.TLRU]. It creates
+// many instances of [lrucore.TLRU] and works based on [LRU].
 // It manages a unified clock for all the separate instances.
-type TLRU[K comparable, V any] struct {
-	cluster[K, V, *lrucore.TTLCore[K, V]]
+type PoolTLRU[K comparable, V any] struct {
+	pool[K, V, *lrucore.TLRU[K, V]]
 	clock *lruclock.Clock
 }
 
@@ -27,7 +27,7 @@ type tlruConfig struct {
 	sliding bool
 }
 
-// TLRUOption is used to configure [TLRU] when creating an instance using [NewTTL] constructor.
+// TLRUOption is used to configure [TLRU] when creating an instance using [NewWithTTL] constructor.
 type TLRUOption interface {
 	apply(c *tlruConfig) error
 }
@@ -45,8 +45,8 @@ func (f LRUOption) apply(c *tlruConfig) error {
 	return f(&c.lruConfig)
 }
 
-// NewTTL creates a [TLRU] instance with the given capacity, ttl and options. It creates
-// the required [lrucore.TTLCore] instances, initiates the [mux.Mux] for shard routing.
+// NewWithTTL creates a [TLRU] instance with the given capacity, ttl and options. It creates
+// the required [lrucore.TLRU] instances, initiates the [mux.Mux] for shard routing.
 // It defaults to the Mux with hash/maphash algorithm. Check `tlru/mux` package for alternatives.
 //
 // The ttl value is rounded off in terms of its internal clock ticks. Check [lruclock.Clock.Ticks].
@@ -58,7 +58,7 @@ func (f LRUOption) apply(c *tlruConfig) error {
 //
 // Returns [ErrInvalidCapacity] if capacity is not in the range of int32
 // and greater than or equal to twice the number of shards.
-func NewTTL[K comparable, V any](capacity int, ttl time.Duration, opts ...TLRUOption) (*TLRU[K, V], error) {
+func NewWithTTL[K comparable, V any](capacity int, ttl time.Duration, opts ...TLRUOption) (*PoolTLRU[K, V], error) {
 	// build the config
 	cfg := tlruConfig{
 		lruConfig: lruConfig{
@@ -93,27 +93,27 @@ func NewTTL[K comparable, V any](capacity int, ttl time.Duration, opts ...TLRUOp
 		_ = cfg.clock.Start()
 	}
 
-	createShard := func(cap int) (*lrucore.TTLCore[K, V], error) {
+	createShard := func(cap int) (*lrucore.TLRU[K, V], error) {
 		if cfg.sliding {
-			return lrucore.NewTTL[K, V](cap, ttl, lrucore.WithClock(cfg.clock), lrucore.WithSliding())
+			return lrucore.NewWithTTL[K, V](cap, ttl, lrucore.WithClock(cfg.clock), lrucore.WithSliding())
 		}
-		return lrucore.NewTTL[K, V](cap, ttl, lrucore.WithClock(cfg.clock))
+		return lrucore.NewWithTTL[K, V](cap, ttl, lrucore.WithClock(cfg.clock))
 	}
-	cluster, err := assemble(capacity, cfg.shards, hash, createShard)
+	pool, err := assemble(capacity, cfg.shards, hash, createShard)
 	if err != nil {
 		return nil, err
 	}
 
-	return &TLRU[K, V]{
-		cluster: cluster,
-		clock:   cfg.clock,
+	return &PoolTLRU[K, V]{
+		pool:  pool,
+		clock: cfg.clock,
 	}, nil
 }
 
-// WithClock allows the usage of a custom clock for [TLRU].
+// WithClock allows the usage of a custom clock for [PoolTLRU].
 // It is only initialized if "TTL" is enabled.
 //
-// NOTE: Using WithClock on [NewTTL] will not start the clock. Use [lruclock.Clock.Start] to
+// NOTE: Using WithClock on [NewWithTTL] will not start the clock. Use [lruclock.Clock.Start] to
 // initiate the timer.
 func WithClock(clock *lruclock.Clock) tlruOpt {
 	return func(c *tlruConfig) error {
@@ -124,8 +124,8 @@ func WithClock(clock *lruclock.Clock) tlruOpt {
 
 // WithSliding enables Sliding TTL on the LRU cache.
 //
-// It will update the timestamp of the key on [TLRU.Get] and
-// [TLRU.Put].
+// It will update the timestamp of the key on [PoolTLRU.Get] and
+// [PoolTLRU.Put].
 func WithSliding() tlruOpt {
 	return func(c *tlruConfig) error {
 		c.sliding = true
@@ -134,64 +134,64 @@ func WithSliding() tlruOpt {
 }
 
 // Close safely closes the background clock when TTL is enabled on the cache.
-func (l *TLRU[K, V]) Close() {
+func (l *PoolTLRU[K, V]) Close() {
 	if l.clock != nil {
 		l.clock.Stop()
 	}
 }
 
-// Check [TLRU.Get] on how Get works.
+// Check [PoolTLRU.Get] on how Get works.
 // It also returns the remaining TTL in the key if it was found in the cache.
-func (l *TLRU[K, V]) GetWithTTL(key K) (V, time.Duration, bool) {
+func (l *PoolTLRU[K, V]) GetWithTTL(key K) (V, time.Duration, bool) {
 	shard := l.mux(key)
-	return l.cluster.shards[shard].GetWithTTL(key)
+	return l.pool.shards[shard].GetWithTTL(key)
 }
 
-// Check [TLRU.Peek] on how Peek works.
+// Check [PoolTLRU.Peek] on how Peek works.
 // It also returns the remaining TTL in the key if it was found in the cache.
-func (l *TLRU[K, V]) PeekWithTTL(key K) (V, time.Duration, bool) {
+func (l *PoolTLRU[K, V]) PeekWithTTL(key K) (V, time.Duration, bool) {
 	shard := l.mux(key)
-	return l.cluster.shards[shard].PeekWithTTL(key)
+	return l.pool.shards[shard].PeekWithTTL(key)
 }
 
 // PutWithTTL adds a new value to the cache with the given key and the provided ttl value.
 //
 // The ttl value is rounded off in terms of its internal clock ticks.
 //
-// Check [TLRU.Put] on how Put works.
-func (l *TLRU[K, V]) PutWithTTL(key K, value V, ttl time.Duration) {
+// Check [PoolTLRU.Put] on how Put works.
+func (l *PoolTLRU[K, V]) PutWithTTL(key K, value V, ttl time.Duration) {
 	shard := l.mux(key)
-	l.cluster.shards[shard].PutWithTTL(key, value, ttl)
+	l.pool.shards[shard].PutWithTTL(key, value, ttl)
 }
 
 // Refresh resets the TTL of an existing key using the default ttl.
 // It returns false if the key could not be found.
-func (l *TLRU[K, V]) Refresh(key K) bool {
+func (l *PoolTLRU[K, V]) Refresh(key K) bool {
 	shard := l.mux(key)
-	return l.cluster.shards[shard].Refresh(key)
+	return l.pool.shards[shard].Refresh(key)
 }
 
 // SetTTL resets the TTL of an existing key using the provided ttl value.
 // It returns false if the key could not be found.
 //
 // The ttl value is rounded off in terms of its internal clock ticks.
-func (l *TLRU[K, V]) SetTTL(key K, ttl time.Duration) bool {
+func (l *PoolTLRU[K, V]) SetTTL(key K, ttl time.Duration) bool {
 	shard := l.mux(key)
-	return l.cluster.shards[shard].SetTTL(key, ttl)
+	return l.pool.shards[shard].SetTTL(key, ttl)
 }
 
 // TTL returns the remaining TTL for the key.
-func (l *TLRU[K, V]) TTL(key K) (time.Duration, bool) {
+func (l *PoolTLRU[K, V]) TTL(key K) (time.Duration, bool) {
 	shard := l.mux(key)
-	return l.cluster.shards[shard].TTL(key)
+	return l.pool.shards[shard].TTL(key)
 }
 
 // UpsertWithTTL adds a new value to the cache with the given key and the provided ttl value.
 //
 // The ttl value is rounded off in terms of its internal clock ticks.
 //
-// Check [TLRU.Upsert] on how Upsert works.
-func (l *TLRU[K, V]) UpsertWithTTL(key K, value V, ttl time.Duration) (lrucore.UpsertState, V) {
+// Check [PoolTLRU.Upsert] on how Upsert works.
+func (l *PoolTLRU[K, V]) UpsertWithTTL(key K, value V, ttl time.Duration) (lrucore.UpsertState, V) {
 	shard := l.mux(key)
-	return l.cluster.shards[shard].UpsertWithTTL(key, value, ttl)
+	return l.pool.shards[shard].UpsertWithTTL(key, value, ttl)
 }